[Del] Remove the related terms tab from term metadata.
Also remove related components. Based on PR feedback.
