+++ src/component/term/DefinitionRelatedTermsEdit.scss
@@ -7,8 +7,3 @@ th.definitionally-related-terms {
 .definitional-occurrence-edit:not(:first-child) {
   margin-left: 1.5rem;
 }
-
-.occurrence-suggested-info {
-  font-style: italic;
-  font-size: 90%;
-}
