+++ src/component/term/TermMetadata.tsx
@@ -1,17 +1,16 @@
 import * as React from "react";
-import { injectIntl } from "react-intl";
-import withI18n, { HasI18n } from "../hoc/withI18n";
-import { Card, CardBody, Col, Row } from "reactstrap";
+import {injectIntl} from "react-intl";
+import withI18n, {HasI18n} from "../hoc/withI18n";
+import {Card, CardBody, Col, Row} from "reactstrap";
 import Term from "../../model/Term";
 import UnmappedProperties from "../genericmetadata/UnmappedProperties";
 import TermAssignments from "./TermAssignments";
 import Tabs from "../misc/Tabs";
-import RelatedTerms from "./RelatedTerms";
 import AssetHistory from "../changetracking/AssetHistory";
 import Terms from "./Terms";
 import BasicTermMetadata from "./BasicTermMetadata";
 import Vocabulary from "../../model/Vocabulary";
-import { RouteComponentProps, withRouter } from "react-router";
+import {RouteComponentProps, withRouter} from "react-router";
 import LanguageSelector from "../multilingual/LanguageSelector";
 import ValidationResults from "./validation/ValidationResults";
 import Utils from "../../util/Utils";
@@ -178,7 +177,6 @@ export class TermMetadata extends React.Component<
     tabs["term.metadata.assignments.title"] = (
       <TermAssignments term={term} onLoad={this.setAssignmentsCount} />
     );
-    tabs["term.metadata.related.tab.title"] = <RelatedTerms term={term} />;
     tabs["history.label"] = <AssetHistory asset={term} />;
     tabs["term.metadata.validation.title"] = <ValidationResults term={term} />;
     tabs["comments.title"] = (
